perf: context-sensitive keywords: for uncore_foo/miss/
Previously the inbuilt event names could not be re-used in subsystem
context, like uncore_foo/miss/, because "miss" was always recognized
as a builtin keyword by the lexer.

Now such useful event names as "cycles", "store", "miss" and "ops"
are only recognized as builtin events in the <INITIAL> state,
so they're free to be re-used without conflict as the names of
events within add-in subsystems. This allows event-naming to follow
hardware docs, avoiding an extra level of name-mapping documentation.

Running "perf stat -a -e cycles,uncore_dma/cycles/" is now allowed.

Signed-off-by: Carlos Munoz <cmunoz@caviumnetworks.com>